+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.44.0](https://github.com/DiluDevX/deliveroo-clone-web/compare/v1.43.0...v1.44.0) (2026-07-18)
+
+### Features
+
+* enhance RestaurantMenuPage with pagination and category dropdown for better navigation ([57c750e](https://github.com/DiluDevX/deliveroo-clone-web/commit/57c750e6a4db64a1692701189fc2294edb81778f))
+
 ## [1.43.0](https://github.com/DiluDevX/deliveroo-clone-web/compare/v1.42.0...v1.43.0) (2026-07-18)
 
 ### Features